requirements.txt
# Core dependencies
openai>=1.12.0

# Optional but recommended
python-dotenv>=1.0.0

# For development/testing (optional)
# pytest>=7.0.0
# black>=22.0.0
# flake8>=4.0.0
